Quick refactoring
