+++ config/config.py
@@ -1,3 +1,2 @@
-ACCEPTED_DML_TYPES = ['INSERT', 'REPLACE', 'UPDATE']
 DEFAULT_SCHEMA = 'dwh'
 DELIMITER = ';;'
