+++ output/cmd.py
@@ -1,5 +1,4 @@
 from colorama import Fore, Style
-from config.config import ACCEPTED_DML_TYPES
 from typing import List
 from sql.objects import Procedure
 
@@ -22,15 +21,15 @@ def beautify(results: List[Procedure]) -> None:
             msg += f'|\n|--- {Fore.BLUE}{q.operation}{Style.RESET_ALL} ' \
                    f'----> {Fore.CYAN}{q.target_table.name}\n{Style.RESET_ALL}'
 
-            if q.operation in ACCEPTED_DML_TYPES:
-                if q.from_table:
-                    for f in q.from_table:
-                        msg += f'{Fore.CYAN}                  ' \
-                               f'. {Fore.BLUE}FROM{Fore.CYAN} {f.name}{Style.RESET_ALL}\n'
-                if q.join_table:
-                    for f in q.join_table:
-                        msg += f'{Fore.CYAN}                  ' \
-                               f'. {Fore.BLUE}JOIN{Fore.CYAN} {f.name}{Style.RESET_ALL}\n'
+            if q.from_table:
+                for f in q.from_table:
+                    msg += f'{Fore.CYAN}                  ' \
+                           f'. {Fore.BLUE}FROM{Fore.CYAN} {f.name}{Style.RESET_ALL}\n'
+            if q.join_table:
+                for f in q.join_table:
+                    msg += f'{Fore.CYAN}                  ' \
+                           f'. {Fore.BLUE}JOIN{Fore.CYAN} {f.name}{Style.RESET_ALL}\n'
+
             if q.target_columns:
                 msg += f'. {Fore.BLUE}Columns --> {Fore.CYAN} {q.target_columns}{Style.RESET_ALL}\n\n'
